// src/domains/dashboard/Payroll/types/salaryProfileTypes/employeeSalaryTable.ts
export type employeeSalaryListingResponse = {
    count: number;
    salaryCycle: SalaryCycle;
    totalPayableSum: number;
    rows: employeeDetails[];
};
export type SalaryCycle = {
    SalaryCycleStart: string;
    SalaryCycleEnd: string;
    SalaryCycleDays: number;
    salaryCycleStart: string;
    salaryCycleEnd: string;
    salaryCycleDays: number;
    workingDays:number;
};
type employeeDetails = {
    totalDeduction: number;
    corporateUser: string;
    employee: Employee;
    year: number;
    month: number;
    salaryCycleStart: string;
    salaryCycleEnd: string;
    salaryCycleDays: number;
    leaveCount: number;
    leaveDeduction: number;
    attendancePercentage: number;
    salaryInformation: SalaryInformation;
    workSchedule: WorkSchedule;
    gratuityContribution: number;
    reimbursements: any[];
    totalReimbursement: number;
    totalPayable: number;
    others: number;
    department: {
        _id: string;
        departmentName: string;
    };
    status: boolean;
    paySlipEmailSent: boolean;
    message: string;
    paymentStatus: string;
    paidViaRollout: boolean;
    createdAt: string;
    updatedAt: string;
    id: string;
    totalOtherDeduction: number;
    totalBonus: number;
    totalOvertime: number;
    totalIncentive: number;
    totalArrears: number;
    monthlySalary: number;
    bankDetails:any[];
};

type EmployeeInformation = {
    dateOfJoin: string;
    employeeId: string;
    designation: string;
    workLocation: string;
    status: string;
    employeeStatus: string;
};

type offBoardingInformation = {
    lastWorkingDay: string;
    resignationLetter: string;
    noticePeriod: number;
    offBoardingType: string;
    reasonForOffBoarding: string;
};

type WorkSchedule = {
    startTime: string;
    endTime: string;
    breakTimeHrs: number;
    days: {
        monday: boolean;
        tuesday: boolean;
        wednesday: boolean;
        thursday: boolean;
        friday: boolean;
        saturday: boolean;
        sunday: boolean;
    };
};

type SalaryInformation = {
    deductionAmount: number | 0;
    leavesAmount: number | 0;
    tdsAmount?: number;
    basicPay: number | 0;
    travelAllowances: number;
    homeAllowances: number;
    medicalAllowances: number;
    otherAllowances: number;
    other: number;
};

type Employee = {
    corporateUser: string;
    profileImage: string;
    personalInformation: {
        fullName: string;
        dateOfBirth: string;
        gender: string;
        mobileNo: string;
        email: string;
        personalAddress: string;
    };
    employeeInformation: EmployeeInformation;
    offBoardingInformation?: offBoardingInformation;
    emergencyNo?: string;
    id: string;
};

export type salarytableType = {
    id: string;
    name: string;
    employeeId: string;
    role: string;
    basicSalary: string | number;
    others: string | number;
    totalSalary?: string;
    totalDeduction?: string | number;
    status: string;
    action: string;
    email: string;
    image: string;
    department: string;
    salaryId: string;
    eId: string;
    employeeStatus?: string;
    lastWorkingDay?: string;
    monthlySalary?: string | number;
    totalPayable?: string | number;
    tds?: string | number;
    netSalaryBeforeTax?: string | number;
    oneTimeAmount?: number;
};